Oberoi Pokhran Road 2 Thane
Oberoi Pokhran Road 2 is an outstanding housing property located in Thane by Oberoi Realty. The venture provides appropriate profits that include the main place, a happy and luxurious lifestyle, great amenities, healthy vicinities, and high recurrence to the residents. It offers 2, 3, and 4 BHK Luxury Apartments s with all types of facilities & accommodations.

Pre-Launch Oberoi Realty Pokhran Road 2 is purposely set and offers direct connectivity to nearly all other major points in and around. It is one of the most reliable addresses in the city with easy access to many famed colleges, leisure areas, shopping areas, hospitals, public gardens, and many other public features.

Oberoi Pokhran Road 2 Thane, It's a well-known group in the real estate market in Maharashtra. The team of this builder group is known for its greater work and on-time delivery of Housing Apartments advanced exactly in harmony with the pre-defined provisions.

Oberoi Flats in Pokhran Road 2 Thane offers apartments of many extents. The size of the area included in this property differs depending on the number of BHKs. The master plan of the scheme is designed in such a way that these studios include wide spaces with proper ventilation at every corner of the home. The interiors are elegantly crafted with a designer tiled floor, granite pawn slab in the kitchen, contemporary clean fittings in the bathroom, and huge windows for proper daylight.

Eases and Facilities: The facilities obtainable in this venture are Indoor Games, Swimming Pool, Gymnasium, a Club House, Wi-Fi, and Safety. All these landscapes together satisfy many major to relax, refresh and enjoy at own home. Apart from that, suitable safety devices are installed to ensure the safety of the residents 24*7 hours.

Home-seekers can also go to decide on updated photo colonnades, floor plans, latest offers, reviews, builder info, and neighborhood info for a well understanding of the development.

For More Details :
Visit - https://oberoi.developerprojects.in/thane/oberoi-pokhran-road-2/